move Monomial and KeyComparable to a separate module
